+++ packages/framework/src/context/request.application.context.ts
@@ -1,29 +1,27 @@
-import { GenericApplicationContext } from "./generic.application.context";
-import type { IApplicationContext } from "../interface";
-import { REQUEST_CTX_KEY } from "../constant";
-import { PIPELINE_IDENTIFIER } from "../decorators/decorator.manager";
+import { GenericApplicationContext } from './generic.application.context'
+import type { IApplicationContext } from '../interface'
+import { REQUEST_CTX_KEY } from '../constant'
+import { PIPELINE_IDENTIFIER } from '../decorators/decorator.manager'
 
 export class RequestApplicationContext extends GenericApplicationContext {
-  private readonly applicationContext: IApplicationContext;
+  private readonly applicationContext: IApplicationContext
 
   constructor(ctx: any, applicationContext: IApplicationContext) {
-    super(applicationContext);
-    this.applicationContext = applicationContext;
+    super(applicationContext)
+    this.applicationContext = applicationContext
 
     // update legacy relationship
-    this.registry.setIdentifierRelation(
-      this.applicationContext.registry.getIdentifierRelation(),
-    );
+    this.registry.setIdentifierRelation(this.applicationContext.registry.getIdentifierRelation())
 
-    this.ctx = ctx;
+    this.ctx = ctx
     // register ctx
-    this.registerObject(REQUEST_CTX_KEY, ctx);
+    this.registerObject(REQUEST_CTX_KEY, ctx)
     // register res
-    this.registerObject("res", {});
+    this.registerObject('res', {})
 
     if (ctx.logger) {
       // register contextLogger
-      this.registerObject("logger", ctx.logger);
+      this.registerObject('logger', ctx.logger)
     }
   }
 
@@ -32,63 +30,59 @@ export class RequestApplicationContext extends GenericApplicationContext {
   }
 
   override get<T = any>(identifier: any, args?: any): T {
-    if (typeof identifier !== "string") {
-      identifier = this.getIdentifier(identifier);
-    }
+    const obj = this.getObject(identifier)
+    if (obj) return obj
 
-    if (this.registry.hasObject(identifier)) {
-      return this.registry.getObject(identifier);
-    }
-
-    const definition =
-      this.applicationContext.registry.getDefinition(identifier);
+    const definition = this.applicationContext.registry.getDefinition(identifier)
     if (definition) {
-      if (
-        definition.isRequestScope() ||
-        definition.id === PIPELINE_IDENTIFIER
-      ) {
+      if (definition.isRequestScope() || definition.id === PIPELINE_IDENTIFIER) {
         // create object from applicationContext definition for requestScope
         return this.getManagedResolverFactory().create({
           definition,
-          args,
-        });
+          args
+        })
       }
     }
 
     if (this.parent) {
-      return this.parent.get(identifier, args);
+      return this.parent.get(identifier, args)
     }
-    return undefined as unknown as T;
+    return undefined as unknown as T
   }
-
-  async getAsync<T = any>(identifier: any, args?: any): Promise<T> {
-    if (typeof identifier !== "string") {
-      identifier = this.getIdentifier(identifier);
+  private getObject<T = any>(identifier: any): T {
+    if (typeof identifier !== 'string') {
+      identifier = this.getIdentifier(identifier)
     }
 
     if (this.registry.hasObject(identifier)) {
-      return this.registry.getObject(identifier);
+      return this.registry.getObject(identifier)
     }
 
-    const definition =
-      this.applicationContext.registry.getDefinition(identifier);
+    // 如果父级有，则直接返回
+    if (this.parent.hasObject(identifier)) {
+      return this.parent.registry.getObject(identifier)
+    }
+    return undefined as unknown as T
+  }
+  async getAsync<T = any>(identifier: any, args?: any): Promise<T> {
+    const obj = this.getObject(identifier)
+    if (obj) return obj
+
+    const definition = this.applicationContext.registry.getDefinition(identifier)
     if (definition) {
-      if (
-        definition.isRequestScope() ||
-        definition.id === PIPELINE_IDENTIFIER
-      ) {
+      if (definition.isRequestScope() || definition.id === PIPELINE_IDENTIFIER) {
         // create object from applicationContext definition for requestScope
         return this.getManagedResolverFactory().createAsync({
           definition,
-          args,
-        });
+          args
+        })
       }
     }
 
     if (this.parent) {
-      return this.parent.getAsync<T>(identifier, args);
+      return this.parent.getAsync<T>(identifier, args)
     }
-    return undefined as unknown as T;
+    return undefined as unknown as T
   }
 
   async ready(): Promise<void> {
@@ -96,6 +90,6 @@ export class RequestApplicationContext extends GenericApplicationContext {
   }
 
   getContext(): any {
-    return this.ctx;
+    return this.ctx
   }
 }