+++ packages/pouchdb/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @electron-boot/pouchdb
 
+## 2.1.11
+
+### Patch Changes
+
+- 优化对象查找
+- Updated dependencies
+  - @electron-boot/framework@2.1.9
+
+## 2.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @electron-boot/framework@2.1.8
+
 ## 2.1.9
 
 ### Patch Changes